+++ deployer/default/deploy/task/deploy_fast.php
@@ -10,6 +10,9 @@ task('deploy-fast', [
     // Read more on https://github.com/sourcebroker/deployer-extended#deploy-check-composer-install
     'deploy:check_composer_install',
 
+    // Read more on https://github.com/sourcebroker/deployer-extended#deploy-check-branch
+    'deploy:check_branch',
+
     // Standard deployer deploy:prepare
     'deploy:prepare',
 
@@ -58,6 +61,10 @@ task('deploy-fast', [
     // Standard deployer cleanup.
     'cleanup',
 
-])->desc('Deploy your TYPO3');
+    // Read more on https://github.com/sourcebroker/deployer-extended#deploy-extend-log
+    'deploy:extend_log',
 
-after('deploy-fast', 'success');
+    // Standard deployer success.
+    'success',
+
+])->desc('Deploy your TYPO3');