+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ master
 
 1) [TASK] Add branch check tasks to deploy.
 2) [TASK][BREAKING] Remove deployer-bulk-tasks dependency. Add database:updateschema task.
+3) [TASK][BREAKING] Update database synchro config for TYPO3 9.
 
 9.0.0
 ~~~~~